The final version of assignment 1
When working on subsequent parts of the compiler, I made these backups of what I turned in, as a sort of rudimentary version control. I did this because there was the threat of immediate failure if my code was found/used publicly during the duration of the class, and I felt more confident in keeping things local than using an unfamiliar private version control.